# Fix performance regression of simple indexing cases