This includes https://github.com/pytorch/pytorch/pull/6770 (with some improvements/bug fixes as suggested in the reviews).