@gchanan 2018-04-20
Benchmark results from the performance regression part on my machine.

Before:
```
a = torch.rand(100, 100, 100)

In [5]: timeit a[[2, 1, 0 ,4]]
The slowest run took 6.93 times longer than the fastest. This could mean that an intermediate result is being cached.
1000 loops, best of 3: 511 µs per loop

In [6]: timeit a[[2, 1, 0 ,4]] = 5
1000 loops, best of 3: 830 µs per loop


In [8]: timeit a[a > 0.5]
10 loops, best of 3: 65.8 ms per loop

In [9]: timeit a[a > 0.5] = 1
10 loops, best of 3: 73.7 ms per loop
```

After:
```
a = torch.rand(100, 100, 100)

In [3]: %timeit a[[2, 1, 0 ,4]]
The slowest run took 37.17 times longer than the fastest. This could mean that an intermediate result is being cached.
10000 loops, best of 3: 34.2 µs per loop

In [4]: %timeit a[[2, 1, 0 ,4]] = 5
The slowest run took 7.63 times longer than the fastest. This could mean that an intermediate result is being cached.
10000 loops, best of 3: 53.5 µs per loop

In [5]: %timeit a[a > 0.5]
10 loops, best of 3: 17.7 ms per loop

In [6]: %timeit a[a > 0.5] = 1
100 loops, best of 3: 15.7 ms per loop
```

Not sure what the slow run warning is about.